refactor: mv cmd subcmd
