+++ main.go
@@ -5,12 +5,12 @@ package main
 import (
 	"os"
 
-	"github.com/vim-volt/volt/cmd"
 	"github.com/vim-volt/volt/logger"
+	"github.com/vim-volt/volt/subcmd"
 )
 
 func main() {
-	err := cmd.Run(os.Args, cmd.DefaultRunner)
+	err := subcmd.Run(os.Args, subcmd.DefaultRunner)
 	if err != nil {
 		logger.Error(err.Msg)
 		os.Exit(err.Code)
