+++ subcmd/build_test.go
@@ -1,4 +1,4 @@
-package cmd
+package subcmd
 
 import (
 	"bytes"
@@ -11,12 +11,12 @@ import (
 	"time"
 
 	"github.com/haya14busa/go-vimlparser"
-	"github.com/vim-volt/volt/cmd/builder"
 	"github.com/vim-volt/volt/config"
 	"github.com/vim-volt/volt/fileutil"
 	"github.com/vim-volt/volt/internal/testutil"
 	"github.com/vim-volt/volt/lockjson"
 	"github.com/vim-volt/volt/pathutil"
+	"github.com/vim-volt/volt/subcmd/builder"
 )
 
 // Checks:
