+++ subcmd/buildinfo/buildinfo.go
