+++ subcmd/help_test.go
@@ -1,4 +1,4 @@
-package cmd
+package subcmd
 
 import (
 	"strings"
